requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
dataclasses-json==0.5.7
Deprecated==1.2.14
duckdb==0.8.0
faiss-cpu==1.7.4
fastapi==0.96.0
Flask==2.3.2
Flask-Limiter==3.3.1
Flask-Session==0.5.0
flatbuffers==23.5.26
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.190
limits==3.5.0
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
ordered-set==4.1.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
posthog==3.0.1
protobuf==4.23.2
pulsar-client==3.2.0
pydantic==1.10.8
Pygments==2.15.1
pypdf==3.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.15.0
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
